Rename to untyped_, ++format
