+++ src/node/governance_history.h
@@ -20,5 +20,6 @@ namespace ccf
   };
   DECLARE_JSON_TYPE(GovernanceHistoryEntry)
   DECLARE_JSON_REQUIRED_FIELDS(GovernanceHistoryEntry, signed_request)
-  using GovernanceHistory = kv::JsonSerialisedMap<MemberId, GovernanceHistoryEntry>;
+  using GovernanceHistory =
+    kv::JsonSerialisedMap<MemberId, GovernanceHistoryEntry>;
 }
